Move MultisignatureExecute and MessageHadnler tests to /service
